Week 10 Security and db

+++ src/main/java/ie/cit/cloudapp/JdbcTodoRepository.java
@@ -4,12 +4,14 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 
-
 import javax.sql.DataSource;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.context.SecurityContextHolder;
 
+@Secured("ROLE_USER")
 public class JdbcTodoRepository {
 
 	private JdbcTemplate jdbcTemplate;
@@ -17,32 +19,39 @@ public class JdbcTodoRepository {
 	public JdbcTodoRepository(DataSource dataSource) {
 		this.jdbcTemplate = new JdbcTemplate(dataSource);
 	}
-	
-	JdbcTodoRepository(){	
+
+	JdbcTodoRepository() {
 	}
 
 	public void save(Todo todo) {
-		jdbcTemplate.update("insert into TODO(text, done) values(?,?)",
-				todo.getText(), todo.isDone());
+		jdbcTemplate.update(
+				"insert into TODO(text, done, owner) values(?,?,?)",
+				todo.getText(), todo.isDone(), getCurrentUser());
 	}
 
 	public Todo get(int id) {
 		return jdbcTemplate.queryForObject(
-				"select id, text, done from TODO where id=?", new TodoMapper(),
-				id);
+				"select id, text, done from TODO where id=? and owner=?", new TodoMapper(),
+				id, getCurrentUser());
 	}
-	
+
 	public List<Todo> getAll() {
-		return jdbcTemplate.query("select id, text, done from TODO", new TodoMapper());
+		return jdbcTemplate.query("select id, text, done from TODO where owner=?",
+				new TodoMapper(), getCurrentUser());
+	}
+
+	private String getCurrentUser() {
+		return SecurityContextHolder
+		.getContext().getAuthentication().getName();
 	}
 
 	public void delete(int id) {
-		jdbcTemplate.update("delete from TODO where id=?", id);
+		jdbcTemplate.update("delete from TODO where id=? and owner=?", id, getCurrentUser());
 	}
 
 	public void update(Todo todo) {
-		jdbcTemplate.update("update TODO set text=?, done=? where id=?",
-				todo.getText(), todo.isDone(), todo.getId());
+		jdbcTemplate.update("update TODO set text=?, done=? where id=? and owner=?",
+				todo.getText(), todo.isDone(), todo.getId(), getCurrentUser());
 	}
 
 }
